% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/get_recipes.R
=======
% Please edit documentation in R/find_recipes.R
>>>>>>> week-3
\name{find_recipes}
\alias{find_recipes}
\title{Get Recipes Based on Ingredients That You Have at Home}
\usage{
<<<<<<< HEAD
find_recipes(ingredients, file)
=======
find_recipes(ingredients, file, max_minutes = Inf)
>>>>>>> week-3
}
\arguments{
\item{ingredients}{A character vector of ingredients you have or want to use.}

\item{file}{Path to the Recipes.csv file.}
<<<<<<< HEAD
=======

\item{max_minutes}{Maximum time (in minutes) to filter recipes. Default is Inf (no filter).}
>>>>>>> week-3
}
\value{
A character vector of recipe names and all of the relevant ingredients
}
\description{
This function suggests recipes that can be made
using your available ingredients at home
or simply ingredients that you want to use.
}
\details{
The recipes are based on the Recipes.csv file.
}
\examples{
<<<<<<< HEAD
get_recipes(c("chicken", "pepper", "flour"),
                     system.file("extdata", "Recipes.csv", package = "recipefinder"))
=======
find_recipes(c("chicken", "pepper", "flour"),
                     system.file("extdata", "Recipes.csv", package = "recipefinder"), 30)
>>>>>>> week-3
}
