% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_list.R
\name{recipes_list}
\alias{recipes_list}
\title{List All Recipes}
\usage{
<<<<<<< HEAD
recipes_list(
  file = system.file("extdata", "Recipes.csv", package = "recipefinder")
)
=======
recipes_list(file)
>>>>>>> week-3
}
\arguments{
\item{file}{Path to the Recipes.csv file. Default is the internal copy in extdata.}
}
\value{
A character vector of all the recipe titles.
}
\description{
Returns a character vector of all recipe titles available in the package's recipe database.
}
\examples{
recipes_list()
}
