% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recipes.R
\name{find_recipes}
\alias{find_recipes}
\title{Get Recipes Based on Ingredients That You Have at Home}
\usage{
find_recipes(ingredients, file)
}
\arguments{
\item{ingredients}{A character vector of ingredients you have or want to use.}

\item{file}{Path to the Recipes.csv file.}
}
\value{
A character vector of recipe names and all of the relevant ingredients
}
\description{
This function suggests recipes that can be made
using your available ingredients at home
or simply ingredients that you want to use.
}
\details{
The recipes are based on the Recipes.csv file.
}
\examples{
get_recipes(c("chicken", "pepper", "flour"),
                     system.file("extdata", "Recipes.csv", package = "recipefinder"))
}
