% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recipes_list.R
\name{recipes_list}
\alias{recipes_list}
\title{List All Recipes}
\usage{
recipes_list(file)
}
\arguments{
\item{file}{Path to the Recipes.csv file. Default is the internal copy in extdata.}
}
\value{
A character vector of all the recipe titles.
}
\description{
Returns a character vector of all recipe titles available in the package's recipe database.
}
\examples{
recipes_list()
}
